requirements.txt
pinecone-client
sentence_transformers